Replace Mfkey32 with MFKey
