+++ base_pack/mfkey/application.fam
@@ -1,19 +1,27 @@
 App(
-    appid="mfkey32",
-    name="Mfkey32",
+    appid="mfkey",
+    name="MFKey",
     apptype=FlipperAppType.EXTERNAL,
     targets=["f7"],
-    entry_point="mfkey32_main",
+    entry_point="mfkey_main",
     requires=[
         "gui",
         "storage",
     ],
     stack_size=1 * 1024,
-    fap_description="Mf Classic key finder",
-    fap_version="1.1",
     fap_icon="mfkey.png",
     fap_category="NFC",
     fap_author="@noproto",
     fap_icon_assets="images",
     fap_weburl="https://github.com/noproto/FlipperMfkey",
+    fap_description="MIFARE Classic key recovery tool",
+    fap_version="2.0",
+)
+
+App(
+    appid="mfkey_init_plugin",
+    apptype=FlipperAppType.PLUGIN,
+    entry_point="init_plugin_ep",
+    requires=["mfkey"],
+    sources=["init_plugin.c"],
 )
