--- base_pack/mfkey32_nested/.gitignore
@@ -1,4 +1,0 @@
-dist/*
-.vscode
-.clang-format
-.editorconfig
