--- base_pack/mfkey32_nested/application.fam
@@ -1,17 +1,0 @@
-App(
-    appid="mfkeynested",
-    name="MFKey Nested",
-    apptype=FlipperAppType.EXTERNAL,
-    targets=["f7"],
-    entry_point="mfkey_main",
-    requires=[
-        "gui",
-        "storage",
-    ],
-    stack_size=1 * 1024,
-    fap_icon="mfkey.png",
-    fap_category="NFC",
-    fap_author="noproto",
-    fap_icon_assets="images",
-    fap_weburl="https://github.com/noproto/FlipperMfkey",
-)
